pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Rheingold Salon - SalonGPT"
authors = ["Nils Imdahl"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.109.2"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
sqlalchemy = "^2.0.26"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
pytest-dotenv = "^0.5.2"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
sse-starlette = "^2.0.0"
boto3 = "^1.0.0"
httpx = {extras = ["http2"], version = "^0.27.0"}
inquirer = "^3.2.4"
py-expression-eval = "^0.3.14"
xmltodict = "^0.13.0"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"
bcrypt = "^4.1.2"
pypdf = "^4.2.0"
pyjwt = "^2.8.0"
pydantic-settings = "^2.3.1"
google_auth_oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.133.0"
openpyxl = "^3.1.5"
kombu = "^5.3.7"
watchdog = "^4.0.1"
redis = {extras = ["hiredis"], version = "^5.0.7"}
python-docx = "^1.1.2"
python-calamine = "^0.2.3"
pyarrow = "^17.0.0"
structlog = "^24.4.0"
pyyaml = "^6.0.1"
nltk = "^3.9.1"
hyperframe = "^6.0.1"
huggingface-hub = "^0.26.5"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
pytest-env = "^1.1.3"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
pre-commit = "^2.20.0"
ruff = "^0.6.0"
pytest-asyncio = "^0.23.7"

[tool.poetry.group.setup]
optional = true

[tool.poetry.group.setup.dependencies]
python = "~3.11"
inquirer = "3.2.4"
python-dotenv = "1.0.1"
cohere = "^5.0.2"
sqlalchemy = "^2.0.26"
fastapi = "^0.109.2"
boto3 = "^1.0.0"
py-expression-eval = "^0.3.14"
langchain = "0.2.15"
langchain-cohere = "^0.1.1"
tavily-python = "^0.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = [
  "src/backend/services/metrics.py", 
  "src/backend/tools/google_drive/sync/actions/",
]
defineConstant = { DEBUG = true }
reportMissingImports = true
reportMissingTypeStubs = false

[tool.ruff]
lint.select = ["F", "I", "E", "W", "I"]
lint.ignore = ["E501", "N806"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
